[refactor] 불필요한 파라미터 제거, 파일 위치 및 이름 조정
- event 객체 전달 받을 필요 없어서 파라미터에서 제거함
- useModalStatus가 다른 모달창과 공통으로 사용되기 힘들었어서, Import창에 맞춰서 사용하기 위해 파일명과 위치 조정함
